Integrate platform initialization in cmd client
